Separate message handling into a ttn layer and a node layer
Both different TTN stacks and different node types are now implemented in orthogonal subclasses.
